Make separate files for precog utilities
